+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## [1.1.0] - 07/31/2024
+## [2.0.0] - 07/31/2024
 
 - Add optional size and scaleFactor parameters
 - Introduce constructors on Gutter for types like small, large, etc